feat: removed members and invited user tables from sso page
