--- frontend/src/container/OrganizationSettings/PendingInvitesContainer/styles.tsx
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-
-export const TitleWrapper = styled.div`
-	display: flex;
-	flex-direction: row;
-	justify-content: space-between;
-	align-items: center;
-`;
